Add Container CRD
Define a build-less build CRD as `containers.build.projectriff.io`. That
will resolve an image to a digest using the same mechanism as
Application and Function.

There are now three developer abstractions:
- Function - build with function buildpacks
- Application - build with application buildpacks
- Container - use a pre built image

Tools that can consume riff build types can point to a Container
resource to pick up new images with the same interface as Application
and Function.