Harmonize Inputs/InputNames and Outputs/OutputNames
Combine the dual slice of strings into a single slice of a binding
struct. This makes the behavior much easier to read as a user and easier
for tools to manipulate, and there is only a single slice to manipulate
rather than two that need to be kept in sync.